+++ R/Artifact.R
@@ -22,11 +22,30 @@ ArtifactRecord <- R6::R6Class( # nolint object_name_linter
       load_file(file_path, suffix, ...)
     },
     #' @description
-    #' Cache the artifact to the local filesystem. This currently only supports
-    #' S3 storage.
+    #' Cache the artifact to the local filesystem. When the Python `lamindb`
+    #' package is not available this only supports S3 storage.
     #'
     #' @return The path to the cached artifact
     cache = function() {
+
+      py_lamin <- private$.instance$get_py_lamin()
+      if (!is.null(py_lamin)) {
+        if (isTRUE(private$.instance$is_default)) {
+          py_artifact <- py_lamin$Artifact$get(self$uid)
+        } else {
+          instance_settings <- private$.instance$get_settings()
+          slug <- paste0(instance_settings$owner, "/", instance_settings$name)
+
+          py_artifact <- py_lamin$Artifact$using(slug)$get(self$uid)
+        }
+        return(py_artifact$cache()$path)
+      }
+
+      cli::cli_warn(paste(
+        "The Python {.pkg lamindb} package is not available.",
+        "Loaded artifacts will not be tracked."
+      ))
+
       # assume that an artifact will have a storage field,
       # and that the storage field will have a type field
       artifact_storage <- private$get_value("storage")