+++ R/Instance.R
@@ -60,22 +60,21 @@ create_instance <- function(instance_settings, is_default = FALSE) {
   )
 
   py_lamin <- NULL
-  if (isTRUE(is_default)) {
-    check_requires("Connecting to Python", "reticulate", alert = "warning")
-
-    py_lamin <- tryCatch(
-      reticulate::import("lamindb"),
-      error = function(err) {
-        cli::cli_warn(c(
-          paste(
-            "Failed to connect to the Python {.pkg lamindb} package,",
-            "you will not be able to create records"
-          ),
-          "i" = "See {.run reticulate::py_config()} for more information"
-        ))
-        NULL
-      }
-    )
+  check_requires("Connecting to Python", "reticulate", alert = "warning")
+  py_lamin <- tryCatch(
+    reticulate::import("lamindb"),
+    error = function(err) {
+      NULL
+    }
+  )
+  if (isTRUE(is_default) && is.null(py_lamin)) {
+    cli::cli_warn(c(
+      paste(
+        "Default instance failed to connect to the Python {.pkg lamindb} package,",
+        "you will not be able to create records"
+      ),
+      "i" = "See {.run reticulate::py_config()} for more information"
+    ))
   }
 
   # create the instance